refactor: move case state to context
