+++ src/app/cases/[id]/ClientCasePage.tsx
@@ -1,511 +1,80 @@
 "use client";
-
-import { apiEventSource, apiFetch } from "@/apiClient";
 import CaseChat from "@/app/cases/[id]/CaseChat";
-import useDragReset from "@/app/cases/useDragReset";
 import CaseLayout from "@/app/components/CaseLayout";
 import CaseProgressGraph from "@/app/components/CaseProgressGraph";
 import DebugWrapper from "@/app/components/DebugWrapper";
-import useCaseAnalysisActive from "@/app/useCaseAnalysisActive";
-import { useSession } from "@/app/useSession";
-import { withBasePath } from "@/basePath";
 import type { Case } from "@/lib/caseStore";
-import {
-  getCaseOwnerContact,
-  getCasePlateNumber,
-  getCasePlateState,
-  getCaseVin,
-  getRepresentativePhoto,
-  hasViolation,
-} from "@/lib/caseUtils";
-import { useRouter } from "next/navigation";
-import { useEffect, useRef, useState } from "react";
-import { useNotify } from "../../components/NotificationProvider";
+import { CaseProvider, useCaseContext } from "./CaseContext";
 import CaseDetails from "./components/CaseDetails";
 import CaseExtraInfo from "./components/CaseExtraInfo";
 import CaseHeader from "./components/CaseHeader";
 import ClaimBanner from "./components/ClaimBanner";
 import PhotoSection from "./components/PhotoSection";
 
-export default function ClientCasePage({
-  initialCase,
+function CasePageContent({
   caseId,
-  initialIsAdmin = false,
-  readOnly = false,
-}: {
-  initialCase: Case | null;
-  caseId: string;
-  initialIsAdmin?: boolean;
-  readOnly?: boolean;
-}) {
-  const [caseData, setCaseData] = useState<Case | null>(initialCase);
-  const analysisActive = useCaseAnalysisActive(
-    caseId,
-    caseData?.public ?? false,
-  );
-  const [preview, setPreview] = useState<string | null>(null);
-  const [selectedPhoto, setSelectedPhoto] = useState<string | null>(
-    initialCase ? getRepresentativePhoto(initialCase) : null,
-  );
-  const [plate, setPlate] = useState<string>(
-    initialCase ? getCasePlateNumber(initialCase) || "" : "",
-  );
-  const [plateState, setPlateState] = useState<string>(
-    initialCase ? getCasePlateState(initialCase) || "" : "",
-  );
-  const [vin, setVin] = useState<string>(
-    initialCase ? getCaseVin(initialCase) || "" : "",
-  );
-  const [note, setNote] = useState<string>(initialCase?.note || "");
-  const [photoNote, setPhotoNote] = useState<string>("");
-  const [members, setMembers] = useState<
-    Array<{
-      userId: string;
-      role: string;
-      name: string | null;
-      email: string | null;
-    }>
-  >([]);
-  const [copied, setCopied] = useState(false);
-  const [reanalyzingPhoto, setReanalyzingPhoto] = useState<string | null>(null);
-  const { data: session } = useSession();
-  const isAdmin =
-    session?.user?.role === "admin" ||
-    session?.user?.role === "superadmin" ||
-    initialIsAdmin;
-  const router = useRouter();
-  const fileInputRef = useRef<HTMLInputElement | null>(null);
-  const [hasCamera, setHasCamera] = useState(false);
-  const [dragging, setDragging] = useState(false);
-  const [hideClaimBanner, setHideClaimBanner] = useState(false);
-  const [chatExpanded, setChatExpanded] = useState(false);
-  const notify = useNotify();
-  const showClaimBanner = Boolean(
-    caseData?.sessionId && !session?.user && !hideClaimBanner,
-  );
-
-  /* -------------------------------------------------------------------- */
-  /*                               EFFECTS                                */
-  /* -------------------------------------------------------------------- */
-
-  useDragReset(() => {
-    setDragging(false);
-  });
-
-  useEffect(() => {
-    void caseId;
-    setHideClaimBanner(false);
-  }, [caseId]);
-
-  useEffect(() => {
-    void caseData?.sessionId;
-    if (!caseData?.sessionId) {
-      setHideClaimBanner(false);
-    }
-  }, [caseData?.sessionId]);
-
-  useEffect(() => {
-    if (
-      "mediaDevices" in navigator &&
-      typeof navigator.mediaDevices.getUserMedia === "function" &&
-      (location.protocol === "https:" || location.hostname === "localhost")
-    ) {
-      setHasCamera(true);
-    }
-  }, []);
-
-  useEffect(() => {
-    const stored = sessionStorage.getItem(`preview-${caseId}`);
-    if (stored) setPreview(stored);
-    apiFetch(`/api/cases/${caseId}`).then(async (res) => {
-      if (res.ok) {
-        const data = (await res.json()) as Case;
-        setCaseData(data);
-      }
-    });
-    apiFetch(`/api/cases/${caseId}/members`).then(async (res) => {
-      if (res.ok) setMembers(await res.json());
-    });
-    const es = apiEventSource("/api/cases/stream");
-    es.onmessage = (e) => {
-      const data = JSON.parse(e.data) as Case & { deleted?: boolean };
-      if (data.id !== caseId) return;
-      if (data.deleted) {
-        setCaseData(null);
-      } else {
-        setCaseData(data);
-        sessionStorage.removeItem(`preview-${caseId}`);
-      }
-    };
-    return () => es.close();
-  }, [caseId]);
-
-  useEffect(() => {
-    if (caseData) {
-      setPlate(getCasePlateNumber(caseData) || "");
-      setPlateState(getCasePlateState(caseData) || "");
-      setVin(getCaseVin(caseData) || "");
-      setNote(caseData.note || "");
-      setSelectedPhoto((prev) => {
-        const all = new Set<string>([
-          ...caseData.photos,
-          ...(caseData.threadImages ?? []).map((img) => img.url),
-        ]);
-        return prev && all.has(prev) ? prev : getRepresentativePhoto(caseData);
-      });
-    }
-  }, [caseData]);
-
-  useEffect(() => {
-    if (caseData && selectedPhoto) {
-      setPhotoNote(caseData.photoNotes?.[selectedPhoto] || "");
-    }
-  }, [caseData, selectedPhoto]);
-
-  useEffect(() => {
-    if (caseData?.analysisStatus !== "pending") {
-      setReanalyzingPhoto(null);
-    }
-  }, [caseData?.analysisStatus]);
-
-  /* -------------------------------------------------------------------- */
-  /*                            API helpers                               */
-  /* -------------------------------------------------------------------- */
-
-  async function uploadFiles(files: FileList) {
-    if (!files || files.length === 0) return;
-    const results = await Promise.all(
-      Array.from(files).map((file) => {
-        const formData = new FormData();
-        formData.append("photo", file);
-        formData.append("caseId", caseId);
-        return apiFetch("/api/upload", {
-          method: "POST",
-          body: formData,
-        });
-      }),
-    );
-    if (results.some((r) => !r.ok)) {
-      notify("Failed to upload one or more files.");
-      return;
-    }
-    const res = await apiFetch(`/api/cases/${caseId}`);
-    if (res.ok) {
-      const data = (await res.json()) as Case;
-      setCaseData(data);
-    } else {
-      notify("Failed to refresh case after upload.");
-    }
-    router.refresh();
-    if (fileInputRef.current) fileInputRef.current.value = "";
-  }
-
-  async function handleUpload(e: React.ChangeEvent<HTMLInputElement>) {
-    const files = e.target.files;
-    if (files) await uploadFiles(files);
-  }
-
-  async function refreshCase() {
-    const res = await apiFetch(`/api/cases/${caseId}`);
-    if (res.ok) {
-      const data = (await res.json()) as Case;
-      setCaseData(data);
-    } else {
-      notify("Failed to refresh case.");
-    }
-  }
-
-  async function updateVehicle(plateNum: string, plateSt: string) {
-    const res = await apiFetch(`/api/cases/${caseId}/override`, {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        vehicle: {
-          licensePlateNumber: plateNum || undefined,
-          licensePlateState: plateSt || undefined,
-        },
-      }),
-    });
-    if (!res.ok) {
-      notify("Failed to update vehicle information.");
-      return;
-    }
-    await refreshCase();
-  }
-
-  async function updatePlateNumber(value: string) {
-    setPlate(value);
-    await updateVehicle(value, plateState);
-  }
-
-  async function updatePlateStateFn(value: string) {
-    setPlateState(value);
-    await updateVehicle(plate, value);
-  }
-
-  async function clearPlateNumber() {
-    setPlate("");
-    await updateVehicle("", plateState);
-  }
-
-  async function clearPlateState() {
-    setPlateState("");
-    await updateVehicle(plate, "");
-  }
-
-  async function updateVinFn(value: string) {
-    setVin(value);
-    const res = await apiFetch(`/api/cases/${caseId}/vin`, {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ vin: value || null }),
-    });
-    if (!res.ok) {
-      notify("Failed to update VIN.");
-      return;
-    }
-    await refreshCase();
-  }
-
-  async function clearVin() {
-    setVin("");
-    const res = await apiFetch(`/api/cases/${caseId}/vin`, {
-      method: "DELETE",
-    });
-    if (!res.ok) {
-      notify("Failed to clear VIN.");
-      return;
-    }
-    await refreshCase();
-  }
-
-  async function updateCaseNoteFn(value: string) {
-    setNote(value);
-    const res = await apiFetch(`/api/cases/${caseId}/note`, {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ note: value || null }),
-    });
-    if (!res.ok) {
-      notify("Failed to update note.");
-      return;
-    }
-    await refreshCase();
-  }
-
-  async function updatePhotoNoteFn(value: string) {
-    if (!selectedPhoto) return;
-    setPhotoNote(value);
-    const res = await apiFetch(`/api/cases/${caseId}/photo-note`, {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ photo: selectedPhoto, note: value || null }),
-    });
-    if (!res.ok) {
-      notify("Failed to update note.");
-      return;
-    }
-    await refreshCase();
-  }
-
-  async function togglePublic() {
-    const res = await apiFetch(`/api/cases/${caseId}/public`, {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ public: !(caseData?.public ?? false) }),
-    });
-    if (!res.ok) {
-      notify("Failed to update visibility.");
-      return;
-    }
-    await refreshCase();
-  }
-
-  async function toggleClosed() {
-    const res = await apiFetch(`/api/cases/${caseId}/closed`, {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ closed: !(caseData?.closed ?? false) }),
-    });
-    if (!res.ok) {
-      notify("Failed to update status.");
-      return;
-    }
-    await refreshCase();
-  }
-
-  async function toggleArchived() {
-    const res = await apiFetch(`/api/cases/${caseId}/archived`, {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ archived: !(caseData?.archived ?? false) }),
-    });
-    if (!res.ok) {
-      notify("Failed to update status.");
-      return;
-    }
-    await refreshCase();
-  }
-
-  async function copyPublicUrl() {
-    const url = `${window.location.origin}${withBasePath(
-      `/public/cases/${caseId}`,
-    )}`;
-    await navigator.clipboard.writeText(url);
-    setCopied(true);
-    setTimeout(() => setCopied(false), 2000);
-  }
-
-  async function reanalyzePhoto(
-    photo: string,
-    detailsEl?: HTMLDetailsElement | null,
-  ) {
-    const url = `/api/cases/${caseId}/reanalyze-photo?photo=${encodeURIComponent(
-      photo,
-    )}`;
-    if (caseData) setCaseData({ ...caseData, analysisStatus: "pending" });
-    setReanalyzingPhoto(photo);
-    const res = await apiFetch(url, { method: "POST" });
-    if (res.ok) {
-      if (detailsEl) {
-        detailsEl.open = false;
-      }
-    } else {
-      notify("Failed to reanalyze photo.");
-    }
-    await refreshCase();
-  }
-
-  async function retryAnalysis() {
-    if (caseData) setCaseData({ ...caseData, analysisStatus: "pending" });
-    const res = await apiFetch(`/api/cases/${caseId}/reanalyze`, {
-      method: "POST",
-    });
-    if (!res.ok) {
-      notify("Failed to retry analysis.");
-    }
-    await refreshCase();
-  }
-
-  async function removePhoto(photo: string) {
-    const delRes = await apiFetch(`/api/cases/${caseId}/photos`, {
-      method: "DELETE",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ photo }),
-    });
-    if (!delRes.ok) {
-      notify("Failed to remove photo.");
-      return;
-    }
-    const res = await apiFetch(`/api/cases/${caseId}`);
-    if (res.ok) {
-      const data = (await res.json()) as Case;
-      setCaseData(data);
-    } else {
-      notify("Failed to refresh case after removing photo.");
-    }
-    router.refresh();
-    const confirmed = window.confirm("Photo removed. Reanalyze this case now?");
-    if (confirmed) {
-      await apiFetch(`/api/cases/${caseId}/reanalyze`, { method: "POST" });
-      router.refresh();
-    }
-  }
-
-  async function refreshMembers() {
-    const res = await apiFetch(`/api/cases/${caseId}/members`);
-    if (res.ok) setMembers(await res.json());
-  }
-
-  async function inviteMember(userId: string) {
-    if (!userId) return;
-    const res = await apiFetch(`/api/cases/${caseId}/invite`, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ userId }),
-    });
-    if (!res.ok) {
-      notify("Failed to invite collaborator.");
-      return;
-    }
-    await refreshMembers();
-  }
-
-  async function removeMember(uid: string) {
-    const res = await apiFetch(`/api/cases/${caseId}/members/${uid}`, {
-      method: "DELETE",
-    });
-    if (!res.ok) {
-      notify("Failed to remove collaborator.");
-      return;
-    }
-    await refreshMembers();
-  }
-
-  /* -------------------------------------------------------------------- */
-  /*                              RENDER                                  */
-  /* -------------------------------------------------------------------- */
+  readOnly,
+}: { caseId: string; readOnly?: boolean }) {
+  const {
+    caseData,
+    showClaimBanner,
+    setHideClaimBanner,
+    chatExpanded,
+    setChatExpanded,
+    dragging,
+    setDragging,
+    handleUpload,
+    fileInputRef,
+    hasCamera,
+    removePhoto,
+    reanalyzingPhoto,
+    requestValue,
+    progress,
+    progressDescription,
+    analysisActive,
+    photoNote,
+    updatePhotoNote,
+    reanalyzePhoto,
+    selectedPhoto,
+    setSelectedPhoto,
+    copied,
+    copyPublicUrl,
+    ownerContact,
+    isAdmin,
+    violationIdentified,
+    isPhotoReanalysis,
+    plateNumberOverridden,
+    plateStateOverridden,
+    vin,
+    vinOverridden,
+    note,
+    updateVin,
+    clearVin,
+    updateNote,
+    updatePlateNumber,
+    updatePlateState,
+    clearPlateNumber,
+    clearPlateState,
+    retryAnalysis,
+    togglePublic,
+    toggleClosed,
+    toggleArchived,
+    members,
+    canManageMembers,
+    inviteMember,
+    removeMember,
+  } = useCaseContext();
 
   if (!caseData) {
     return (
       <div className="p-8 flex flex-col gap-4">
         <h1 className="text-xl font-semibold">Uploading...</h1>
-        {preview ? (
-          // eslint-disable-next-line @next/next/no-img-element
-          <img src={preview} alt="preview" className="max-w-full" />
-        ) : null}
         <p className="text-sm text-gray-500 dark:text-gray-400">
           Uploading photo...
         </p>
       </div>
     );
   }
-
-  const violationIdentified =
-    caseData.analysisStatus === "complete" && hasViolation(caseData.analysis);
-
-  const vinOverridden = caseData.vinOverride !== null;
-  const plateNumberOverridden =
-    caseData.analysisOverrides?.vehicle?.licensePlateNumber !== undefined;
-  const plateStateOverridden =
-    caseData.analysisOverrides?.vehicle?.licensePlateState !== undefined;
-  const ownerContact = getCaseOwnerContact(caseData);
-  const isOwner = members.some(
-    (m) => m.userId === session?.user?.id && m.role === "owner",
-  );
-  const canManageMembers = isAdmin || isOwner;
-
-  const progress =
-    caseData.analysisStatus === "pending" && caseData.analysisProgress
-      ? caseData.analysisProgress
-      : null;
-  const isPhotoReanalysis = Boolean(
-    reanalyzingPhoto && caseData.analysisStatus === "pending",
-  );
-  const requestValue = progress
-    ? progress.stage === "upload"
-      ? progress.index > 0
-        ? (progress.index / progress.total) * 100
-        : undefined
-      : Math.min((progress.received / progress.total) * 100, 100)
-    : undefined;
-  const progressDescription = progress
-    ? `${progress.steps ? `Step ${progress.step} of ${progress.steps}: ` : ""}${
-        progress.stage === "upload"
-          ? progress.index > 0
-            ? `Uploading ${progress.index} of ${progress.total} photos (${Math.floor(
-                (progress.index / progress.total) * 100,
-              )}%)`
-            : "Uploading photos..."
-          : progress.done
-            ? "Processing results..."
-            : `Analyzing... ${progress.received} of ${progress.total} tokens`
-      }`
-    : caseData.analysisStatus === "pending"
-      ? "Analyzing photo..."
-      : caseData.analysisStatus === "canceled"
-        ? "Analysis canceled."
-        : "Analysis failed.";
 
   return (
     <div
@@ -523,9 +92,8 @@ export default function ClientCasePage({
         readOnly
           ? undefined
           : (e) => {
-              if (!e.currentTarget.contains(e.relatedTarget as Node)) {
+              if (!e.currentTarget.contains(e.relatedTarget as Node))
                 setDragging(false);
-              }
             }
       }
       onDrop={
@@ -533,7 +101,9 @@ export default function ClientCasePage({
           ? undefined
           : async (e) => {
               e.preventDefault();
-              await uploadFiles(e.dataTransfer.files);
+              await handleUpload(
+                e as unknown as React.ChangeEvent<HTMLInputElement>,
+              );
               setDragging(false);
             }
       }
@@ -549,83 +119,18 @@ export default function ClientCasePage({
         }
       >
         <CaseLayout
-          header={
-            <CaseHeader
-              caseId={caseId}
-              caseData={caseData}
-              ownerContact={ownerContact}
-              isAdmin={isAdmin}
-              readOnly={readOnly}
-              violationIdentified={violationIdentified}
-              progress={progress}
-              isPhotoReanalysis={isPhotoReanalysis}
-              copyPublicUrl={copyPublicUrl}
-              copied={copied}
-            />
-          }
+          header={<CaseHeader caseId={caseId} readOnly={readOnly} />}
           left={<CaseProgressGraph caseData={caseData} />}
           right={
             <>
               <DebugWrapper data={caseData}>
-                <CaseDetails
-                  caseData={caseData}
-                  progress={progress}
-                  readOnly={readOnly}
-                  ownerContact={ownerContact}
-                  vin={vin}
-                  vinOverridden={vinOverridden}
-                  note={note}
-                  plateNumberOverridden={plateNumberOverridden}
-                  plateStateOverridden={plateStateOverridden}
-                  updateVin={updateVinFn}
-                  clearVin={clearVin}
-                  updateNote={updateCaseNoteFn}
-                  updatePlateNumber={updatePlateNumber}
-                  updatePlateState={updatePlateStateFn}
-                  clearPlateNumber={clearPlateNumber}
-                  clearPlateState={clearPlateState}
-                  retryAnalysis={retryAnalysis}
-                  canTogglePublic={(isAdmin || session?.user) && !readOnly}
-                  canToggleStatus={(isAdmin || isOwner) && !readOnly}
-                  togglePublic={togglePublic}
-                  toggleClosed={toggleClosed}
-                  toggleArchived={toggleArchived}
-                  members={members}
-                  canManageMembers={canManageMembers}
-                  inviteMember={inviteMember}
-                  removeMember={removeMember}
-                />
+                <CaseDetails readOnly={readOnly} />
               </DebugWrapper>
-
-              <PhotoSection
-                caseId={caseId}
-                caseData={caseData}
-                selectedPhoto={selectedPhoto}
-                setSelectedPhoto={setSelectedPhoto}
-                handleUpload={handleUpload}
-                fileInputRef={fileInputRef}
-                hasCamera={hasCamera}
-                removePhoto={removePhoto}
-                readOnly={readOnly}
-                isPhotoReanalysis={isPhotoReanalysis}
-                reanalyzingPhoto={reanalyzingPhoto}
-                requestValue={requestValue}
-                progress={progress}
-                progressDescription={progressDescription}
-                analysisActive={analysisActive}
-                photoNote={photoNote}
-                updatePhotoNote={updatePhotoNoteFn}
-                reanalyzePhoto={reanalyzePhoto}
-              />
+              <PhotoSection caseId={caseId} readOnly={readOnly} />
             </>
           }
         >
-          <CaseExtraInfo
-            caseId={caseId}
-            caseData={caseData}
-            selectedPhoto={selectedPhoto}
-            setSelectedPhoto={setSelectedPhoto}
-          />
+          <CaseExtraInfo caseId={caseId} />
         </CaseLayout>
       </div>
       {readOnly || !dragging ? null : (
@@ -645,5 +150,28 @@ export default function ClientCasePage({
         />
       </div>
     </div>
+  );
+}
+
+export default function ClientCasePage({
+  initialCase,
+  caseId,
+  initialIsAdmin = false,
+  readOnly = false,
+}: {
+  initialCase: Case | null;
+  caseId: string;
+  initialIsAdmin?: boolean;
+  readOnly?: boolean;
+}) {
+  return (
+    <CaseProvider
+      initialCase={initialCase}
+      caseId={caseId}
+      initialIsAdmin={initialIsAdmin}
+      readOnly={readOnly}
+    >
+      <CasePageContent caseId={caseId} readOnly={readOnly} />
+    </CaseProvider>
   );
 }
